# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API P7",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://api.p7.fr/v1",
            "description": "Mon api"
        }
    ],
    "tags": [
        {
            "tags": "phone"
        },
        {
            "tags": "client"
        },
        {
            "tags": "authentification"
        }
    ],
    "paths": {
        "/api/login_check": {
            "post": {
                "tags": [
                    {
                        "tags": "authentification"
                    }
                ],
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": "string",
                                    "password": "string"
                                },
                                "example":{
                                    "username": "entrez votre email",
                                    "password": "entrez votre password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Votre token JWT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "token": "string"
                                    },
                                    "example":{
                                        "token": "Incluez le dans le header pour effectuer vos requêtes => key: Authorization, value: BEARER votre token ici"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },    
        "/api/clients/{id}": {
            "get": {
                "tags": [
                    {
                        "tags": "client"
                    }
                ],
                "operationId": "App\\Controller\\ClientController::showAction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Le client",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "La ressource n'existe pas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    {
                        "tags": "client"
                    }
                ],
                "operationId": "App\\Controller\\ClientController::updateAction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Client"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Un client",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    {
                        "tags": "client"
                    }
                ],
                "operationId": "App\\Controller\\ClientController::deleteAction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Un client",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/clients": {
            "get": {
                "tags": [
                    {
                        "tags": "client"
                    }
                ],
                "operationId": "App\\Controller\\ClientController::listAction",
                "responses": {
                    "200": {
                        "description": "Les clients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    {
                        "tags": "client"
                    }
                ],
                "operationId": "App\\Controller\\ClientController::createAction",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Client"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Un client",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/phones/{id}": {
            "get": {
                "tags": [
                    {
                        "tags": "phone"
                    }
                ],
                "operationId": "App\\Controller\\PhoneController::showAction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du téléphone",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Le téléphone",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Phone"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "La ressource n'existe pas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    {
                        "tags": "phone"
                    }
                ],
                "operationId": "App\\Controller\\PhoneController::updateAction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du téléphone",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Phone"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Un téléphone",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Phone"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    {
                        "tags": "phone"
                    }
                ],
                "operationId": "App\\Controller\\PhoneController::deleteAction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du téléphone",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Un téléphone",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Phone"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/phones": {
            "get": {
                "tags": [
                    {
                        "tags": "phone"
                    }
                ],
                "operationId": "App\\Controller\\PhoneController::listAction",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "La page a consulter",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nos téléphones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Phone"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    {
                        "tags": "phone"
                    }
                ],
                "operationId": "App\\Controller\\PhoneController::createAction",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Phone"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Un téléphone",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Phone"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Client": {
                "properties": {
                    "nom": {
                        "type": "string"
                    },
                    "prenom": {
                        "type": "string"
                    },
                    "mail": {
                        "type": "string"
                    },
                    "adresse": {
                        "type": "string"
                    },
                    "telephone": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Phone": {
                "properties": {
                    "nom": {
                        "type": "string"
                    },
                    "reference": {
                        "type": "string"
                    },
                    "couleur": {
                        "type": "string"
                    },
                    "dimension": {
                        "type": "string"
                    },
                    "prix": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}